requirements.txt
flask
Pillow
opencv-python-headless
numpy
pytesseract
